requirements.txt
python-telegram-bot==22.5
python-dotenv
requests